Changed collisions & added screen object
Object

Removed the object centering stuff from _object class, so now when you render you send it
the cordinates it will render it from the top-left.

Set the default xy positions of a object to 0.

Player

Added _player::stopX(int x) and _player::stopY(int y) which set the x or y of the player to the one received,
and sets the corresponding velocity to 0.

_player::update doesnt handle collisions with the edge of the screen
anymore. That is done by passing the rect of the screen to collision and true.

changed _player::collision to also receive true or false depending on
whether you want the player to be inside the object and not get out or
if you want the player to be outside and not get in.

Source/Main

Added an object to be the screen.

Made a call for player1.collision to have the player stay inside
the screen object.